test: Adopt tutorial changes in test_tutorial.py
Add, remove and update function calls to match code snippets in
tutorial.

This commit also adds tests for outputs of `repo.status()` and
`repo.dirty_roles()` functions.
Note that the compare-to strings need to be constructed
programatically, akin to how they are constructed in the relevant
functions, in order to avoid issues with unicode prefixes in
Python2/3, e.g.
    "Dirty roles: ['root']"
vs  "Dirty roles: [u'root']"

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>